feat: make app adaptive
