+++ src/index.html
@@ -169,6 +169,5 @@
 </div>
 
 <script src="./assets/js/bootstrap.min.js"></script>
-<script src="./index.ts"></script>
 </body>
 </html>
